optimize for cost after all best matches are found in the batch. this still represents a temporary regression in score versus the naive approach where very high scores for cost efficiency offset the total failure to not exceed the deviation budget. next step: use remaining deviation budget to optimize for cost on most-used inks first.

+++ Ink.h
@@ -12,11 +12,11 @@
 
 class Ink {
 public:
-    Ink(const std::string& color_, float cost_): color(color_), cost(cost_) {}
+    Ink(const std::string& color_, double cost_): color(color_), cost(cost_) {}
     Ink(const Ink& other) : color(other.color), cost(other.cost) {}
     Ink(): color("FFFFFF") {}
     Color color;
-    float cost;
+    double cost;
 };
